using file-based sqlite, added repository & progLang controllers to be able to create/edit/delete these models

+++ server/models/extractionSkillFinding/extractionSkillFindingModel.ts
@@ -12,7 +12,7 @@ export default class ExtractionSkillFindingModel extends Model<InferAttributes<E
     @AutoIncrement
     declare id: CreationOptional<number>
 
-    @Attribute(DataTypes.DECIMAL)
+    @Attribute(DataTypes.DOUBLE)
     @NotNull
     declare score: number
